ACCUMULO-94: make createTableBatchWriter and createMultiTableBatchWriter consistent
git-svn-id: https://svn.apache.org/repos/asf/incubator/accumulo/branches/1.3@1190420 13f79535-47bb-0310-9956-ffa450edef68

+++ src/core/src/main/java/org/apache/accumulo/core/client/mock/MockConnector.java
@@ -60,7 +60,7 @@ public class MockConnector extends Connector {
   }
   
   @Override
-  public MultiTableBatchWriter createMultiTableBatchWriter(long maxMemory, int maxLatency, int maxWriteThreads) {
+  public MultiTableBatchWriter createMultiTableBatchWriter(long maxMemory, long maxLatency, int maxWriteThreads) {
     return null;
   }
